pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "traffic-control-env"
version = "1.0.0"
description = "Autonomous Traffic Signal Control OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Tarun Singh"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "openai>=1.12.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["traffic_env", "traffic_env.server", "server"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

